Chatbot/requirements.txt
# requirements
streamlit
openai
python-dotenv